@@ -18,81 +15,82 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/platform-server": "5.2.9",
-    "@angular/router": "5.2.9",
-    "@angular/service-worker": "5.2.9",
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/platform-server": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@angular/service-worker": "6.0.3",
     "@cloudinary/angular-5.x": "1.0.1",
-    "@fortawesome/angular-fontawesome": "0.1.0-8",
-    "@fortawesome/fontawesome-svg-core": "1.2.0-10",
-    "@fortawesome/free-brands-svg-icons": "5.1.0-4",
-    "@fortawesome/free-solid-svg-icons": "5.1.0-4",
-    "@fortawesome/pro-regular-svg-icons": "5.1.0-4",
-    "@ng-bootstrap/ng-bootstrap": "1.1.1",
+    "@fortawesome/angular-fontawesome": "0.1.0-10",
+    "@fortawesome/fontawesome-svg-core": "1.2.0-11",
+    "@fortawesome/free-brands-svg-icons": "5.1.0-11",
+    "@fortawesome/free-solid-svg-icons": "5.1.0-11",
+    "@fortawesome/pro-regular-svg-icons": "5.1.0-11",
+    "@ng-bootstrap/ng-bootstrap": "2.0.0",
     "@ngx-share/button": "5.3.1",
     "@ngx-share/buttons": "5.3.1",
     "@ngx-share/core": "5.3.1",
-    "@ngx-translate/core": "9.1.1",
-    "@progress/kendo-angular-buttons": "4.0.0",
-    "@progress/kendo-angular-charts": "2.0.3",
-    "@progress/kendo-angular-dateinputs": "2.2.0",
-    "@progress/kendo-angular-dialog": "3.4.0",
-    "@progress/kendo-angular-dropdowns": "2.1.0",
-    "@progress/kendo-angular-excel-export": "1.0.7",
-    "@progress/kendo-angular-grid": "3.0.0",
-    "@progress/kendo-angular-inputs": "2.2.0",
-    "@progress/kendo-angular-intl": "1.4.0",
-    "@progress/kendo-angular-l10n": "1.1.0",
-    "@progress/kendo-angular-upload": "3.0.3",
-    "@progress/kendo-data-query": "1.2.2",
-    "@progress/kendo-drawing": "1.5.2",
-    "@progress/kendo-theme-bootstrap": "2.11.10",
-    "@types/mandrill-api": "1.0.30",
-    "angular2-moment": "1.8.0",
-    "angularfire2": "5.0.0-rc.6",
-    "bootstrap": "4.0.0",
+    "@ngx-translate/core": "10.0.2",
+    "@progress/kendo-angular-buttons": "4.1.1",
+    "@progress/kendo-angular-charts": "3.2.0",
+    "@progress/kendo-angular-dateinputs": "3.4.0",
+    "@progress/kendo-angular-dialog": "3.6.1",
+    "@progress/kendo-angular-dropdowns": "3.0.1",
+    "@progress/kendo-angular-excel-export": "2.1.0",
+    "@progress/kendo-angular-grid": "3.4.0",
+    "@progress/kendo-angular-inputs": "3.1.2",
+    "@progress/kendo-angular-intl": "1.4.1",
+    "@progress/kendo-angular-l10n": "1.2.0",
+    "@progress/kendo-angular-upload": "4.1.2",
+    "@progress/kendo-data-query": "1.3.1",
+    "@progress/kendo-drawing": "1.5.6",
+    "@progress/kendo-theme-bootstrap": "2.13.5",
+    "angular2-moment": "1.9.0",
+    "angularfire2": "5.0.0-rc.10",
+    "bootstrap": "4.1.1",
     "cloudinary-core": "2.5.0",
-    "core-js": "2.5.5",
-    "firebase": "4.12.1",
+    "core-js": "2.5.7",
+    "firebase": "5.0.4",
     "hammerjs": "2.0.8",
-    "mandrill-api": "1.0.45",
-    "moment": "^2.22.0",
+    "moment": "2.22.2",
     "ng2-validators": "2.2.0",
-    "ngx-clipboard": "10.0.0",
+    "ngx-clipboard": "11.1.0",
     "npm-font-open-sans": "1.1.0",
-    "rxjs": "5.5.8",
-    "soltsice": "0.11.4",
+    "openzeppelin-solidity": "1.9.0",
+    "rxjs": "6.2.0",
+    "rxjs-compat": "^6.2.0",
+    "soltsice": "0.11.5",
+    "tslint-angular": "^1.1.2",
     "typeface-raleway": "0.0.54",
-    "web3": "1.0.0-beta.33",
-    "zeppelin-solidity": "1.8.0",
+    "web3": "1.0.0-beta.34",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.9",
-    "@types/jasmine": "2.8.6",
+    "@angular/cli": "6.0.7",
+    "@angular/compiler-cli": "6.0.3",
+    "@types/jasmine": "2.8.8",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "9.6.2",
-    "codelyzer": "4.2.1",
+    "@types/node": "10.1.2",
+    "codelyzer": "4.3.0",
     "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "2.0.2",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "1.0.0",
-    "protractor": "5.3.1",
-    "truffle-contract": "3.0.4",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.6.2"
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.1.0",
+    "protractor": "5.3.2",
+    "truffle-contract": "3.0.5",
+    "ts-node": "6.1.0",
+    "tslint": "5.10.0",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
   }
 }
