@@ -1,17 +1,16 @@
 import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, CanLoad, Route, RouterStateSnapshot} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {UserSessionService} from './user-session.service';
 import {AngularFireAuth} from 'angularfire2/auth';
 import {User} from '../model/user';
+import {map} from 'rxjs/operators';
 
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class HansGuard implements CanActivate, CanLoad {
     private isHans(url: string): Observable<boolean> {
-        return this.userSessionService.getUser()
-            .map((user: User) => {
-                return user.isHans;
-            });
+        return this.userSessionService.getUser().pipe(
+            map((user: User) => user.isHans));
     }
 
     // ===== IMPLEMENTED INTERFACES =====
