@@ -4,10 +4,10 @@
 
 import {Injectable} from '@angular/core';
 import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {environment} from '../../environments/environment';
 
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class MediaTypeHttpInterceptorService implements HttpInterceptor {
 
   private isApiUrl(url): boolean {
