@@ -4,12 +4,12 @@
 
 import {Injectable} from '@angular/core';
 import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {environment} from '../../environments/environment';
 
 const TIMEZONE_OFFSET = 'TimeZone-Offset';
 
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class TimezoneOffsetHttpInterceptorService implements HttpInterceptor {
 
   private isApiUrl(url): boolean {
