@@ -2,16 +2,17 @@
  * Copyright (c) 2017. Traveliko PTE.LTD. All rights Reserved.
  */
 
-import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {SwUpdate} from '@angular/service-worker';
-import 'rxjs/add/observable/interval';
 
-@Injectable()
+import {interval as observableInterval, Observable} from 'rxjs';
+import {Injectable} from '@angular/core';
+import {SwUpdate} from '@angular/service-worker';
+
+
+@Injectable({providedIn: 'root'})
 export class CheckForUpdateService {
 
   constructor(updates: SwUpdate) {
-    Observable.interval(6 * 60 * 60).subscribe(() => updates.checkForUpdate());
+    observableInterval(6 * 60 * 60).subscribe(() => updates.checkForUpdate());
   }
 
 }
