@@ -5,7 +5,7 @@
 import {Injectable} from '@angular/core';
 import {SwUpdate} from '@angular/service-worker';
 
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class LogUpdateService {
 
   constructor(updates: SwUpdate) {
