@@ -3,7 +3,7 @@ function _window(): any {
   // return the global native browser window object
   return window;
 }
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class WindowRefService {
   get nativeWindow(): any {
     return _window();
